% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{prefix_query}
\alias{prefix_query}
\alias{build_sparql}
\title{Perform}
\usage{
prefix_query(query)

build_sparql(query)
}
\arguments{
\item{query}{custom query}
}
\description{
Function to create custom queries with the `/landregistry/query endpoint`. All
necessary prefixes have to be included and output will be parsed from json to
dataframe.
}
