% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_query}
\alias{get_query}
\title{Get the sparql query performed with \code{sparql}.}
\usage{
get_query(x)
}
\arguments{
\item{x}{the result of query.}
}
\description{
Get the sparql query performed with \code{sparql}.
}
