% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail-ons.R
\name{ons_countries}
\alias{ons_countries}
\alias{ons_eng_regions}
\alias{ons_regions}
\alias{ons_eng_counties}
\alias{ons_la}
\title{Office of National Statistic Location Classification}
\usage{
ons_countries(modify = TRUE)

ons_eng_regions(modify = TRUE)

ons_regions(modify = TRUE)

ons_eng_counties(modify = TRUE)

ons_la(modify = FALSE)
}
\arguments{
\item{modify}{Modifies the vector to conform with the queries.}
}
\value{
A character vector with the location names.
}
\description{
Functions \code{ons_countries}, \code{ons_regions}, \code{ons_eng_counties}
and \code{ons_la} provide a vector of countries, regions, English counties
and local authorities respectively.
}
\examples{
\donttest{
ons_countries()
}
}
