% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail_ons.R
\name{ons_lookup}
\alias{ons_lookup}
\alias{ons_countries}
\alias{ons_regions}
\alias{ons_eng_counties}
\alias{ons_la}
\title{Office of National Statistic Location Classification}
\usage{
ons_lookup()

ons_countries(modify = TRUE)

ons_regions(modify = TRUE)

ons_eng_counties(modify = TRUE)

ons_la(modify = FALSE)
}
\arguments{
\item{modify}{Modifies the vector to conformity with the package.}
}
\description{
\code{ons_lookup} provides a table that that combines Local Authority
District 1 and 2 to NUTS3, NUTS2 and NUTS1. Fucntion \code{ons_countries},
\code{ons_regions}, \code{ons_eng_counties} and \code{ons_la}
provide a vector of countries, regions, English counties and local authorities
respectively.
}
\examples{

ons_lookup()


}
