% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail-ons.R
\name{ons_pc}
\alias{ons_pc}
\title{UK Postcodes}
\usage{
ons_pc(pattern = NULL)
}
\arguments{
\item{regexp}{partial matching}
}
\description{
Vector of UK postcodes. Search for matches with the `pattern` argument.
}
\details{
ons_pc("EH21 8A")
}
