% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{retrieve_query}
\alias{retrieve_query}
\title{Get the sparql query performed with \code{sparql}.}
\usage{
retrieve_query(x)
}
\arguments{
\item{x}{the result of query.}
}
\value{
Returns the a character vector with the query.
}
\description{
Get the sparql query performed with \code{sparql}.
}
