% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail-ukhp.R
\name{ukhp_avail_categories}
\alias{ukhp_avail_categories}
\alias{ukhp_avail_regions}
\title{Display Land Registry available categories and regions}
\usage{
ukhp_avail_categories()

ukhp_avail_regions()
}
\description{
Display Land Registry available categories and regions
}
