% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail-ukhp.R
\name{ukhp_avail_items}
\alias{ukhp_avail_items}
\alias{ukhp_avail_regions}
\alias{ukhp_avail_date_span}
\alias{ukhp_avail_date_last}
\title{Display Land Registry available categories}
\usage{
ukhp_avail_items()

ukhp_avail_regions()

ukhp_avail_date_span()

ukhp_avail_date_last()
}
\value{
Returns a character vector.
}
\description{
\code{ukhp_avail_regions()} displays available regions, \code{ukhp_avail_regions()}
displays available items, \code{ukhp_avail_date_span()} displays available date sample,
and \code{ukhp_avail_date_last} displays the last available date.
}
\examples{

ukhp_avail_regions()[1:10]
ukhp_avail_items()
ukhp_avail_date_last()

}
