% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{uklr_browse}
\alias{uklr_browse}
\alias{ukhp_browse}
\alias{ukppd_browse}
\alias{uktrans_browse}
\title{Quickly browse to Land Registry's webpage}
\usage{
uklr_browse()

ukhp_browse()

ukppd_browse()

uktrans_browse()
}
\description{
These functions take you to land registry's webpages and return the URL invisibly.
}
\examples{
uklr_browse()

}
