% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{uklr_def}
\alias{uklr_def}
\alias{ukhp_def}
\alias{ukppd_def}
\alias{uktrans_def}
\title{Linked Data Definitions}
\usage{
uklr_def()

ukhp_def()

ukppd_def()

uktrans_def()
}
\value{
Returns a tibble.
}
\description{
These functions return the definitions for the linked data construction in
a tibble.
}
\examples{
\donttest{
uklr_def()

ukhp_def()
}
}
