% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail-ons.R
\docType{data}
\name{uklr_pc}
\alias{uklr_pc}
\title{UK Postcodes and NUTS3 Codes}
\format{
A tibble with 1,759,911 rows and 2 variables:
\describe{
  \item{NUTS318CD}{NUTS3 code}
  \item{PC}{Postcode}
}
}
\source{
\url{http://ec.europa.eu/eurostat/tercet/}
}
\usage{
uklr_pc
}
\description{
A dataset containing UK postcodes and their corresponding NUTS3 codes.
}
\examples{
\dontrun{
head(uklr_pc)
}
}
\keyword{datasets}
