% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail-ukppd.R
\name{ukppd_avail_categories}
\alias{ukppd_avail_categories}
\alias{ukppd_avail_optional_categories}
\title{Display Price Paid Date available categories}
\usage{
ukppd_avail_categories()

ukppd_avail_optional_categories()
}
\description{
Display Price Paid Date available categories
}
