% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail-ukppd.R
\name{ukppd_avail_items}
\alias{ukppd_avail_items}
\alias{ukppd_avail_optional_items}
\alias{ukppd_avail_postcodes}
\title{Display Price Paid Date available categories}
\usage{
ukppd_avail_items()

ukppd_avail_optional_items()

ukppd_avail_postcodes()
}
\description{
Display Price Paid Date available categories
}
