% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{ukppd_get}
\alias{ukppd_get}
\title{Get Price Paid Date}
\usage{
ukppd_get(postcode = "PL6 8RU", item = NULL, optional_item = NULL,
  start_date = NULL, end_date = NULL)
}
\description{
Get Price Paid Date
}
\examples{

ukppd_get("PL6 8RU")
ukppd_get("PL6 8RU", start_date = "2001-01-01")
ukppd_get("PL6 8RU", start_date = "2001-01-01")
}
