% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{uktrans_get}
\alias{uktrans_get}
\title{Get Transaction Data}
\usage{
uktrans_get(item = NULL, region = NULL, regexp = TRUE,
  start_date = NULL, end_date = NULL, ...)
}
\arguments{
\item{item}{item to select}

\item{region}{the region to select. If regexp is set to FALSE then the matching
should be exact, see \code{ukhp_avail_regions} for available regions. If regexp
is set to TRUE then partial matching is possible. Furthermore it is set to
NULL selects all available regions.}

\item{regexp}{use regular expression in sparql to search for regions.}

\item{start_date}{the start date}

\item{end_date}{the end date}

\item{...}{query modifiers passed through \code{rdf_modifiers}.}
}
\description{
Get Transaction Data
}
\examples{
\donttest{
uktrans_get(item = "totalApplicationCountByRegion", region = "East Anglia")

# If `region` is left as NULL then it returns all available regions
uktrans_get(item = "totalApplicationCountByRegion")

# Quering all available transaction data
uktrans_get(item = uktrans_avail_items())
}
}
