% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{build_sparql}
\alias{build_sparql}
\alias{prefix_query}
\title{Tools to create a custom SPARQL query}
\usage{
build_sparql(query)

prefix_query(query)
}
\arguments{
\item{query}{custom query}
}
\description{
Function to create custom queries with the `/landregistry/query endpoint`. All
necessary prefixes have to be included and output will be parsed from json to
dataframe. The most common prefixes can be include with \code{prefix_query}
}
\examples{
custom_query <- "select * where {
  <http://landregistry.data.gov.uk/data/ukhpi/region/newport/month/2013-10> ?property ?value
}"
build_sparql(prefix_query(custom_query))
}
