% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{build_sparql}
\alias{build_sparql}
\title{Build custom sparql query with}
\usage{
build_sparql(query)
}
\arguments{
\item{query}{custom query}
}
\description{
Function to create custom queries with the `/landregistry/query endpoint`. All
necessary prefixes have been loaded and output will be parsed from json to
dataframe.
}
