% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail-ons.R
\name{ons_lookup}
\alias{ons_lookup}
\alias{ons_nuts1}
\alias{ons_nuts2}
\alias{ons_nuts3}
\alias{ons_lad}
\title{Office of National Statistic Location Classification}
\usage{
ons_lookup()

ons_nuts1()

ons_nuts2()

ons_nuts3()

ons_lad()
}
\value{
A character vector with the location names.
}
\description{
\code{ons_lookup} provides a table that that combines local authority
districts (lad), local administrative units (lau), nomenclature of territorial
units for statistics(NUTS) 1,2 and 3. The other functions can extract the
mentioned variables.
}
\examples{
\donttest{
ons_lookup()
}
}
