% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail-ons.R
\docType{data}
\name{ons_pc}
\alias{ons_pc}
\alias{pc}
\title{UK Postcodes}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1759911 rows and 2 columns.}
\usage{
ons_pc(pmatch = FALSE)

pc
}
\arguments{
\item{pmatch}{partial matching}
}
\description{
UK Postcodes
}
\keyword{datasets}
