% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{ukhp_get}
\alias{ukhp_get}
\title{Get House Priced data}
\usage{
ukhp_get(region = "newport", item = NULL, start_date = NULL,
  end_date = NULL)
}
\arguments{
\item{region}{the region to select}

\item{item}{the item to select}

\item{start_date}{the start date}

\item{end_date}{the end date}
}
\description{
Get House Priced data
}
