% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail-ukppd.R
\name{ukppd_avail_items}
\alias{ukppd_avail_items}
\alias{ukppd_avail_optional_items}
\alias{ukppd_avail_postcodes}
\title{Display Price Paid Date available categories}
\usage{
ukppd_avail_items()

ukppd_avail_optional_items()

ukppd_avail_postcodes()
}
\value{
Returns a character vector.
}
\description{
\code{ukppd_avail_items()} displays all available items for search,
\code{ukppd_avail_optional_items()} displays optional items that refer to the
location of the transaction and \code{ukppd_avail_postcodes} displays the
available postcodes.
/
}
\examples{

ukppd_avail_items()

ukppd_avail_optional_items()

ukppd_avail_postcodes()[1:10]

}
