% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail-uktrans.R
\name{uktrans_avail_items}
\alias{uktrans_avail_items}
\alias{uktrans_avail_regions}
\title{Display Land Registry Transaction available categories}
\usage{
uktrans_avail_items()

uktrans_avail_regions()
}
\value{
Returns a character vector.
}
\description{
\code{uktrans_avail_items()} and \code{uktrans_avail_regions()} display
available items and regions for search respectively.
}
\examples{

uktrans_avail_items()
uktrans_avail_regions()
}
